+++ prr/core/TariffPlan.java
@@ -5,9 +5,9 @@ import java.io.Serializable;
 public abstract class TariffPlan implements Serializable {
   //private String _name;
 
-  protected abstract double computeCost(Client cl, TEXT com);
+  protected abstract long computeCost(Client cl, TEXT com);
 
-  protected abstract double computeCost(Client cl, VOICE com);
+  protected abstract long computeCost(Client cl, VOICE com);
 
-  protected abstract double computeCost(Client cl, VIDEO com);
+  protected abstract long computeCost(Client cl, VIDEO com);
 }
